FLUOMETER_STEPPER5.ino
This is a contribution to the original project of "Biodesign of the Real World" www.biodesign.cc using bioreporters and the arsenic (now mercury) detection prototype https://biodesign.cc/2015/06/01/scattering-test/. In addition to the detection of fluorescence of the vials containing genetically modified E.Coli that reacts to the concentration of mercury on water, this code activates a system of pumps. These pumps generate the direct proportional amount of black ink droplets that there are in form of contamination inside the vials. It translates the amount of turbidity and fluorescence (therefore contamination of heavy metals) into black bits made out of ink.